long --> int conversion
Ed altre modifiche per la recover connection

+++ CLIENT/GUI/editorWindow/gui_editor.h
@@ -18,24 +18,24 @@ class GUI_Editor : public QWidget
 {
     Q_OBJECT
 public:
-    long documentId;
+    int documentId;
     GIMPdocs *gimpParent;
     GUI_ToolsBar *childToolsBar;
     GUI_UsersBar *childUsersBar;
     GUI_MyTextEdit *childMyTextEdit;
-    QMap<long, QColor*> userColorMap;
+    QMap<int, QColor*> userColorMap;
     //indica se il testo è colorato coi colori degli utenti per identificarli
     bool usersColors;
 
-    explicit GUI_Editor(QWidget *parent, long documentId);
+    explicit GUI_Editor(QWidget *parent, int documentId);
     ~GUI_Editor();
     inline static QString getObjectName(){ return "GUI_Editor";}
     //lo scopo di queste 2 funzioni è di venire chiamate da un più basso livello quando viene aggiunto o rimosso un nuovo utente che sta lavorando allo stesso document
     //si occupano sia del cursore che dell'icona che del colore
-    void addUserToEditorGUI(long userid);
-    void removeUserFromEditorGUI(long userid);
-    void addContributorToCurrentDocument(long userid);
-    void removeContributorFromCurrentDocument(long userid);
+    void addUserToEditorGUI(int userid);
+    void removeUserFromEditorGUI(int userid);
+    void addContributorToCurrentDocument(int userid);
+    void removeContributorFromCurrentDocument(int userid);
 
     //mi serve perchè non posso fare le connect direttamente nel costruttore. Quando sono nel costruttore, la ui2 non è ancora stata caricata quindi la connect va fatta in un secondo momento
     void connectMenuBarActions();
@@ -62,8 +62,8 @@ private:
     Ui::GUI_Editor *ui;
     GUI_ColorsManager colorsManager;
 
-    QColor *getUserColor(long userId);
-    void forgetUserColor(long userId);
+    QColor *getUserColor(int userId);
+    void forgetUserColor(int userId);
 
 signals:
     void menuTools_event(menuTools code);
